The original project did not have src and application (package)
so I made a new project with these files and started working in it. My
plan was to add this project to the repository but you can only have one
project per repository. I think I fixed the problem by moving the
src/application + all the work I did into the original project and
removing the original fxml, main, and application files (which are now
contained in the folders I added).