Work on the controller class.
"Finished" the calculations for the water, step, and expenses habits.
They work very generally in that the data and goal are set to the
correct numbers. Calculating the grade sets it to 0 every time. Need to
add some validation but I want to wait until the calculation actually
works.

calculating the dailyscore needs work